adjust policy details screen methods
