+++ x-pack/test_serverless/functional/test_suites/security/cypress/screens/endpoint_management/policy_details.ts
@@ -6,7 +6,29 @@
  */
 
 import { APP_POLICIES_PATH } from '@kbn/security-solution-plugin/common/constants';
+import { UserAuthzAccessLevel } from './types';
+import { getNoPrivilegesPage } from './common';
 
 export const visitPolicyDetails = (policyId: string): Cypress.Chainable => {
   return cy.visit(`${APP_POLICIES_PATH}/${policyId}`);
+};
+
+export const ensurePolicyDetailsPageAuthzAccess = (
+  policyId: string,
+  accessLevel: UserAuthzAccessLevel,
+  visitPage: boolean = false
+): Cypress.Chainable => {
+  if (visitPage) {
+    visitPolicyDetails(policyId);
+  }
+
+  if (accessLevel === 'none') {
+    return getNoPrivilegesPage().should('exist');
+  }
+
+  if (accessLevel === 'read') {
+    return cy.getByTestSubj('policyDetailsSaveButton').should('not.exist');
+  }
+
+  return cy.getByTestSubj('policyDetailsSaveButton').should('exist');
 };
