+++ x-pack/test_serverless/functional/test_suites/security/cypress/screens/endpoint_management/policy_list.ts
@@ -17,10 +17,6 @@ export const visitPolicyList = (): Cypress.Chainable => {
   return cy.visit(pageById.policyList);
 };
 
-export const visitPolicyDetails = (policyId: string): Cypress.Chainable => {
-  return cy.visit(pageById.policyList + `/${policyId}`);
-};
-
 export const ensurePolicyListPageAuthzAccess = (
   accessLevel: UserAuthzAccessLevel,
   visitPage: boolean = false
@@ -35,24 +31,4 @@ export const ensurePolicyListPageAuthzAccess = (
 
   // Read and All currently are the same
   return getNoPrivilegesPage().should('not.exist');
-};
-
-export const ensurePolicyDetailsPageAuthzAccess = (
-  policyId: string,
-  accessLevel: UserAuthzAccessLevel,
-  visitPage: boolean = false
-): Cypress.Chainable => {
-  if (visitPage) {
-    visitPolicyDetails(policyId);
-  }
-
-  if (accessLevel === 'none') {
-    return getNoPrivilegesPage().should('exist');
-  }
-
-  if (accessLevel === 'read') {
-    return cy.getByTestSubj('policyDetailsSaveButton').should('not.exist');
-  }
-
-  return cy.getByTestSubj('policyDetailsSaveButton').should('exist');
 };
